Fix mangling problems, improve sandbox driver
In, out and uniform variables no longer participate in name mangling, as
they participate in shader linkage.

+++ include/frontend/jl/ast.h
@@ -161,7 +161,8 @@ struct VarDecl : public Decl {
                      MaybeScopeType scope, NodeId initializer = NodeId::null_id(),
                      bool is_builtin = false)
         : Decl{location, NodeKind::VarDecl, identifier,
-               static_cast<uint8_t>(scope) | static_cast<uint8_t>(is_builtin << 7)},
+               static_cast<uint8_t>(static_cast<uint8_t>(scope) |
+                                    static_cast<uint8_t>(is_builtin << 7))},
           annot_type{annot_type},
           initializer{initializer} {}
